start writing Mixer class to record SVs during scrambling
